backend/requirements.txt
Django
djangorestframework
psycopg2-binary
celery
redis
python-dotenv